Moved qtip listeners into a behavior
